Updates to unfolding structure
